+++ src/core/standard/lib.zig
@@ -3,6 +3,7 @@ pub const task = @import("task.zig");
 pub const luau = @import("luau.zig");
 pub const net = @import("net/lib.zig");
 pub const stdio = @import("stdio.zig");
+pub const regex = @import("regex.zig");
 pub const serde = @import("serde/lib.zig");
 pub const crypto = @import("crypto/lib.zig");
 pub const process = @import("process.zig");